Revert "Highlight code client-side (#2786)"
This reverts commit 0d615e3888ec8ab5803a765a784d6c0c2ff644ea.
